+++ NEWS.md
@@ -122,6 +122,9 @@
 
 - The error of `logf` has been improved from 5 ULPs to 3 ULPs.
 
+- `assert` doesn't use `fprintf` anymore, but a custom, light-weight function
+  instead in order to diagnose a failed assertion.
+
 - avr-gcc v15 and up issues a diagnostic for `__int24` and `__uint24` when
   `-pedantic` or similar options are on.  Hence `__extension__` was
   added when using these types in `avr/pgmspace.h`.